OAK-110: NPE in KernelNodeStoreBranch.diffToJsop
It's better if we consolidate a transient add/remove directly to a no-op
already in the NodeStateBuilder instead of adding a special case handling
for that in the resulting NodeState.

Also added a more optimized way of snapshotting the change set in common
NodeStateBuilder cases where no or just a single property or child node
have been modified.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1341833 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 6090307faadae3b3786f9fb50117c86276147b5b