OAK-1285: QueryEngine#executeQuery takes NamePathMapper which is not part of oak-api
Replace the NamePathMapper argument with the set of local namespace
mappings, which is used to construct a NamePathMapper instance for
the query.

Pass that instance directly as a constructor argument to QueryImpl instead
of having the setter in the Query interface.

Replace Name(Path)Mapper.hasSessionLocalMappings() with .getSessionLocalMappings()
so that we can access those mappings in UserQueryManager.

Add NO_BINDINGS and NO_MAPPINGS constants to QueryEngine to make null
arguments more readable.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1560446 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 1b28c974b012b221063f5d0feec4b028083435d1