OAK-1332: Large number of changes to the same node can fill observation queue
Postpone generation of reordering events form the diff of the grandparent
node to the parent node. This fixes the handling of reorderings that
happen directly below the root (of the observed subtree) and optimizes
the common case where there are no reorderings. It also reduces the size
of the event queue by only putting reorderings below a single parent to
the queue at a time.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1561758 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 7fee5a96c2c1401eaeebb1425962b9b2e138a98f