OAK-1645 - Route find queries to Mongo secondary in MongoDocumentStore
The find and query method now explicitly define the read preference to be used. The preference is determined by type of Collection , maxCacheAge and modified time of parent. if preference is secondary then actual preference used is picked from DB level settings

Also exposed a setting 'maxReplicationLagMillis' which determines the max possible lag in replication beyond which its safe to route the read queries to secondaries

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1594161 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 3737ce324620c190a648353c92419e36ff78a1c5