OAK-2278 - Make current testcases pass with new index implementation
Refactor the testcase to use LucenePropertyIndex.  Currently testLucene3 is ignored due to OAK-2283 as type based exclusion is currently not supported for property index

LuceneIndexHelper - Allow null for async. This is only done for testcase to ensure lucene index can be used in sync mode. For production code it should be set to 'async'. Hence @Nonnull annotation is retained

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1640870 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: ce8dbb291a371760dcd144369e497f2f27be7936