OAK-3689 - OakOSGiRepositoryFactory shutting down the repository twice
Fixed another case where if shutdown is invoked multiple times (due to some host application issue) then it would trigger a wait on latch which would not be countdown and shutdown logic would wait for timeout.

Now the shutdown call would maintain a state to determine if shutdown is initiated. If not already done it would initiate the shutdown otherwise the call would be ignored

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1717177 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 0624e70c0e08fc21abe20bae9b5f08befa9f5c3b