OAK-4705: Fulltext parser doesn't allow stand-alone hyphen in search expression
Allow hyphen to be anywhere. If it comes before a word/quoted-word, then it acts as a not for the term that follows it.
Also, it seemed that hyphen followed by quoted term wasn't being parse correctly (the quotes didn't fulfil their purpose) - fixed that
Added a bunch of parsing tests

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1760168 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: cdc4d1adfd25a5f3bd13799e364ba1f6ca3ee390