OAK-4412 - Lucene hybrid index
Refactor the TimedRefreshPolicy to extra policy for sync index i.e. RefreshOnWritePolicy which does the refresh upon write itself without waiting for any refresh interval

Changed the NRTIndex to reuse the reader and make use of openIfChanged to ensure that readers are not created fresh if no change has happened. This fact is later used in IndexNode to avoid refreshing the IndexSearcher. This is an extra defence in addition to update flag in policy implementation.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1760868 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 1c3fbc8a28a8a19800e47600199de4b6f9898dd6