OAK-6271 - Support for importing index files
Add mapping for name of folder on file system to the JCR name in repository.
As the JCR name can have chars which are not suitable for FS we strip such chars
and thus ':data' is stored as 'data' on file system.

This mapping would be recorded in index-details.txt and would be later
used while importing the content back

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1800584 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: b24bac9467f5e74826f279f65947ba276d45beda