OAK-6450 - Stop relying on the service.pid property in SecurityProviderRegistration
Use the oak.component.name component property if the service.pid
is not available. The SecurityProviderRegistration property name is
unchanged, for backwards compatibility reasons.

The objectClass property may not be used as it points to the service
name(s) under which the component is registered. The component.name
property was considered and discarded as it is specific to DS.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1803820 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: bf043a13ad784f2ac37653896ebbca31ab8c8d6b