OAK-6890 - Fix SafeRunnable rethrow policy
SafeRunnable should never rethrow the Throwable it catches. While it might not
make sense in some cases - e.g. OutOfMemoryError - it prevents the SafeRunnable
from being unscheduled from the ScheduledExecutorService that runs it
periodically. Each SafeRunnable represents a background task that is important
for the health of the FileStore and should never be accidentally unscheduled.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1814475 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: ca24f057a9f689396ff1771ec57c59dd59aa7dae