OAK-6966 - Always use TarFiles#size in the estimation phase
The estimation phase should always use the same source for the repository size.
If the sources are different, and their semantics and accuracy differ, the
estimation phase might not be able to correctly compute. Since TarFiles#size is
used when writing the gc.log, it should also be used also when computing the
repository size at the beginning of the estimation phase.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1815894 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: f9d3dfafbed4dc5f903f9f350894402e9b91c968