OAK-7158 - Prevent users from changing the number of retained generations
The number of retained generations should not be set to a value different than
two. Doing otherwiswe could cause data loss due to the way cleanup works. To
prevent misconfigurations, the OSGi and the MBean interfaces prevent users from
setting the number of retained generations. If users try to change the number
of retained generations, a warning message is printed and the number of
retained generations is kept unchanged.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1821240 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: fe3c5f524881e5e626f70b44202976cf631559f9