OAK-7434 - Extract compaction implementations in separate components
Introduce CompactionStrategy, which represents a possible way to carry on the
compaction phase in the scope of the bigger garbage collection process. The
full and tail compaction phases have been extracted to FullCompactionStrategy
and TailCompactionStrategy respectively. Moreover, in order to have full
compaction act as a fallback for tail compaction when a base state can't be
found, CompactionResult has been extended with a new result type and
FallbackCompactionStrategy has been introduced.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1829894 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 9d4b6f3c1678795259b32686b0040d80cfe04693