OAK-7918 - Remove ConsistencyChecker
ConsistencyChecker acted as a middleman between the Check command and
ConsistencyCheckerTemplate. It didn't provide any useful functionality outside
of the context of the Check command. This commit merges ConsistencyChcker in
the Check command and removes. Additionally, this commit moves the tests for
the Check command into the o.a.j.o.segment.tool package.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1847567 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 28f4de09001c1836ae46000bb7318a4bff5f13ca