OAK-8552: Add data store config to skip exists check on download URI generation
This change adds a new optional configuration parameter for AzureDataStore and S3DataStore.  The new parameter, named "presignedHttpDownloadURIVerifyExists", has a default value of true which means the createHttpDownloadURI() method will ensure the requested blob exists in blob storage before creating the signed download URI.  If this parameter is set to false, the exists check will be skipped, resulting in higher signed download URI generation performance but potentially returning a URI to a nonexistent blob in some cases (see OAK-7998).


git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1866044 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 966cfedd8e065658c779993239d7730d4e85c5dd