OAK-8600: Change log level to DEBUG in getRecord() when blob does not exist
With recent optimizations to getRecord() in the cloud data stores, it is
not an unexpected thing for the corresponding blob to not exist. We
want to avoid filling logs with INFO level log messages and stack
traces for missing records if the calling code has a reasonable
expectation that the record may not exist.  A DataStoreException will
still be thrown that must be caught by the caller, in case this
situation is not expected.  Just the log level is changed.


git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1866563 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 21b8ec85045c00db0dc956a8849bb7eadfca4d52